pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pclink"
version = "0.1.0"  # Fresh start version
description = "A secure remote PC control and file management application"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "PCLink Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "PySide6>=6.5.0",
    "psutil>=5.9.0",
    "pyperclip>=1.8.2",
    "mss>=9.0.0",
    "keyboard>=0.13.5",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "cryptography>=41.0.0",
    "qrcode>=7.4.0",
    "pyautogui>=0.9.54",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "nuitka>=1.5.0",
    "pyinstaller>=5.13.0",
    "pre-commit>=2.0",
    "build",
    "svglib>=1.5.1",
    "reportlab>=4.0.0",
]

[project.scripts]
pclink = "pclink.main:main"

[project.urls]
Homepage = "https://github.com/BYTEDz/pclink"
Repository = "https://github.com/BYTEDz/pclink"
Issues = "https://github.com/BYTEDz/pclink/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pclink = ["assets/*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".git", ".hg", ".mypy_cache", ".tox", ".venv", "_build", "buck-out", "build", "dist"]

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = [".git", ".hg", ".mypy_cache", ".tox", ".venv", "_build", "buck-out", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"